Fix post script to make SSL certificate
This package makes "localhost" certificate if it doesn't exist at
package installation but it may fail in a container because of lack of
some /proc files related to a physical machine device.

This fixes the post install script to check such /proc files existence
and if these exists, append them to a random value generator source.